Add files via upload
This file contains two methods for loan approval prediction first one using Machine learning algorithms like Random forest, logistic regression and decision trees. This model is created for the customer using a streamlit platform which gives a web appliation. Where customer can get to know if their loans will be approved or not. It can be also used by banks which will tell in seconds where to approve the loan or not. It takes into consideration the cibil score of the client, loan term, loan amount, asset values, Income per annum and loan amount finally it gives the result "approved" or "not approved".